requirements.txt
torch
torchaudio
torchvision
numpy
pandas
pyarrow
scikit-learn
matplotlib
librosa
soundfile
tqdm
pyyaml
captum
lime
scipy
resampy
requests
beautifulsoup4
Pillow
